requirements.txt
requests
pandas
mojang
schedule
python-dotenv
